// js_modules/ui-core/src/configeditor/types/ConfigEditorUtils.types.ts
export type ConfigEditorRunConfigSchemaFragment = {
  __typename: 'RunConfigSchema';
  rootDefaultYaml: string;
  rootConfigType:
    | {__typename: 'ArrayConfigType'; key: string}
    | {__typename: 'CompositeConfigType'; key: string}
    | {__typename: 'EnumConfigType'; key: string}
    | {__typename: 'MapConfigType'; key: string}
    | {__typename: 'NullableConfigType'; key: string}
    | {__typename: 'RegularConfigType'; key: string}
    | {__typename: 'ScalarUnionConfigType'; key: string};
  allConfigTypes: Array<
    | {
        __typename: 'ArrayConfigType';
        key: string;
        description: string | null;
        isSelector: boolean;
        typeParamKeys: Array<string>;
      }
    | {
        __typename: 'CompositeConfigType';
        key: string;
        description: string | null;
        isSelector: boolean;
        typeParamKeys: Array<string>;
        fields: Array<{
          __typename: 'ConfigTypeField';
          name: string;
          description: string | null;
          isRequired: boolean;
          configTypeKey: string;
          defaultValueAsJson: string | null;
        }>;
      }
    | {
        __typename: 'EnumConfigType';
        givenName: string;
        key: string;
        description: string | null;
        isSelector: boolean;
        typeParamKeys: Array<string>;
        values: Array<{__typename: 'EnumConfigValue'; value: string; description: string | null}>;
      }
    | {
        __typename: 'MapConfigType';
        keyLabelName: string | null;
        key: string;
        description: string | null;
        isSelector: boolean;
        typeParamKeys: Array<string>;
      }
    | {
        __typename: 'NullableConfigType';
        key: string;
        description: string | null;
        isSelector: boolean;
        typeParamKeys: Array<string>;
      }
    | {
        __typename: 'RegularConfigType';
        givenName: string;
        key: string;
        description: string | null;
        isSelector: boolean;
        typeParamKeys: Array<string>;
      }
    | {
        __typename: 'ScalarUnionConfigType';
        key: string;
        scalarTypeKey: string;
        nonScalarTypeKey: string;
        description: string | null;
        isSelector: boolean;
        typeParamKeys: Array<string>;
      }
  >;
};
// ...
export type CompositeConfigTypeForSchemaFragment = {
  __typename: 'CompositeConfigType';
  fields: Array<{
    __typename: 'ConfigTypeField';
    name: string;
    description: string | null;
    isRequired: boolean;
    configTypeKey: string;
    defaultValueAsJson: string | null;
  }>;
};